requirements.txt
matplotlib==3.5.2
numpy==1.21.5
pandas==1.4.4
pymssql==2.2.5
python_docx==0.8.11
scikit_learn==1.0.2
scipy==1.9.1
seaborn==0.11.2
tftb==0.1.4
torch==1.13.1
vmdpy==0.2
